+++ packages/adminui-vue/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+  - user menu `Settings` now contains a sub-menu `Scanner settings` what shows a scanner settings form.
+    Ext based `ubm_desktop-userSettings` form (was shown on setting menu click before these changes) is removed.
 
 ### Deprecated